LPS-176997 - rename modules by SF
