--- modules/dxp/apps/osb/osb-analytics/portal-security-audit-header-test/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal Security Audit Header Test
-Bundle-SymbolicName: com.liferay.portal.security.audit.header.test
-Bundle-Version: 1.0.0
