--- modules/dxp/apps/osb/osb-analytics/portal-security-audit-header/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal Security Audit Header
-Bundle-SymbolicName: com.liferay.portal.security.audit.header
-Bundle-Version: 1.0.0
