--- modules/dxp/apps/osb/osb-analytics/portal-security-audit-header/src/main/java/com/liferay/portal/security/audit/header/internal/configuration/AuditHeaderConfiguration.java
@@ -1,49 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.security.audit.header.internal.configuration;
-
-import aQute.bnd.annotation.metatype.Meta;
-
-import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
-
-/**
- * @author Alvaro Saugar
- */
-@ExtendedObjectClassDefinition(category = "audit")
-@Meta.OCD(
-	id = "com.liferay.portal.security.audit.header.internal.configuration.AuditHeaderConfiguration",
-	localization = "content/Language", name = "audit-header-configuration-name"
-)
-public interface AuditHeaderConfiguration {
-
-	@Meta.AD(
-		deflt = "false", description = "audit-header-malu-help",
-		name = "audit-header-malu-enabled", required = false
-	)
-	public boolean enabledMALU();
-
-	@Meta.AD(
-		deflt = "false", description = "audit-header-apv-help",
-		name = "audit-header-apv-enabled", required = false
-	)
-	public boolean enabledAPV();
-
-	@Meta.AD(
-		deflt = "false", description = "audit-header-scope-help",
-		name = "audit-header-scope-enabled", required = false
-	)
-	public boolean enabledScope();
-
-}
